Updates PIG Gradle Plugin for publication
